refactor: MenuProductFixture 분리
